requirements.txt
python-magic==0.4.18
python-slugify==4.0.1
text-unidecode==1.3
